Scripts for oushudb and clickhouse benchmark test

+++ clickhouse-ddl-float.sql
@@ -31,7 +31,7 @@ create table part
     p_type        String,
     p_size        integer,
     p_container   String,
-    p_retailprice Decimal,
+    p_retailprice float,
     p_comment     String
 ) engine = MergeTree order by tuple();
 
@@ -42,7 +42,7 @@ create table supplier
     s_address   String,
     s_nationkey integer,
     s_phone     String,
-    s_acctbal   Decimal,
+    s_acctbal   float,
     s_comment   String
 ) engine = MergeTree order by tuple();
 
@@ -51,7 +51,7 @@ create table partsupp
     ps_partkey    integer,
     ps_suppkey    integer,
     ps_availqty   integer,
-    ps_supplycost Decimal,
+    ps_supplycost float,
     ps_comment    String
 ) engine = MergeTree order by tuple();
 
@@ -62,7 +62,7 @@ create table customer
     c_address    String,
     c_nationkey  integer,
     c_phone      String,
-    c_acctbal    Decimal,
+    c_acctbal    float,
     c_mktsegment String,
     c_comment    String
 ) engine = MergeTree order by tuple();
@@ -72,7 +72,7 @@ create table orders
     o_orderkey      Int64,
     o_custkey       integer,
     o_orderstatus   char,
-    o_totalprice    Decimal,
+    o_totalprice    float,
     o_orderdate     date,
     o_orderpriority String,
     o_clerk         String,
@@ -87,10 +87,10 @@ create table lineitem
     l_partkey       integer,
     l_suppkey       integer,
     l_linenumber    integer,
-    l_quantity      Decimal,
-    l_extendedprice Decimal,
-    l_discount      Decimal,
-    l_tax           Decimal,
+    l_quantity      float,
+    l_extendedprice float,
+    l_discount      float,
+    l_tax           float,
     l_returnflag    char,
     l_linestatus    char,
     l_shipdate      date,
